updated spell system to use new input manager
