+++ Assets/ScriptableObjects/InputManager.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fe9aa25dba7844e1f8272d492fdc1f00
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
